codementor/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai>=0.8.0
httpx>=0.28.1
python-dotenv==1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
